@@ -1,6 +1,6 @@
 {
     "name": "angular-auth-oidc-client",
-    "version": "4.1.1",
+    "version": "6.0.0",
     "description": "An OpenID Connect Implicit Flow client for Angular",
     "main": "./bundles/angular-auth-oidc-client.umd.js",
     "module": "./modules/angular-auth-oidc-client.es5.js",
@@ -34,29 +34,29 @@
     ],
     "license": "MIT",
     "peerDependencies": {
-        "@angular/core": ">= 4.3.0",
-        "@angular/common": ">= 4.3.0",
-        "@angular/router": ">= 4.3.0"
+        "@angular/core": ">= 6.0.0",
+        "@angular/common": ">= 6.0.0",
+        "@angular/router": ">= 6.0.0"
     },
     "dependencies": {
-        "jsrsasign": "8.0.3"
+        "jsrsasign": "8.0.12"
     },
     "devDependencies": {
-        "@angular/animations": "4.3.0",
-        "@angular/common": "4.3.0",
-        "@angular/compiler": "4.3.0",
-        "@angular/compiler-cli": "4.3.0",
-        "@angular/core": "4.3.0",
-        "@angular/platform-browser": "4.3.0",
-        "@angular/platform-browser-dynamic": "4.3.0",
-        "@angular/platform-server": "4.3.0",
-        "@angular/router": "4.3.0",
+        "@angular/animations": "6.0.0",
+        "@angular/common": "6.0.0",
+        "@angular/compiler": "6.0.0",
+        "@angular/compiler-cli": "6.0.0",
+        "@angular/core": "6.0.0",
+        "@angular/platform-browser": "6.0.0",
+        "@angular/platform-browser-dynamic": "6.0.0",
+        "@angular/platform-server": "6.0.0",
+        "@angular/router": "6.0.0",
         "@compodoc/compodoc": "1.0.0-beta.9",
         "@types/jasmine": "2.6.2",
         "@types/node": "8.0.47",
         "chalk": "1.1.3",
         "codelyzer": "3.0.1",
-        "core-js": "2.4.1",
+        "core-js": "2.5.5",
         "jasmine-core": "2.6.2",
         "karma": "1.7.0",
         "karma-chrome-launcher": "2.1.1",
@@ -69,7 +69,7 @@
         "rollup": "0.41.6",
         "rollup-plugin-commonjs": "8.0.2",
         "rollup-plugin-node-resolve": "3.0.0",
-        "rxjs": "5.5.10",
+        "rxjs": "6.1.0",
         "shelljs": "0.7.7",
         "sorcery": "0.10.0",
         "ts-loader": "2.1.0",
@@ -77,6 +77,6 @@
         "typescript": "2.7.2",
         "webpack": "3.8.1",
         "yargs": "8.0.1",
-        "zone.js": "0.8.11"
+        "zone.js": "0.8.26"
     }
 }
